Remove dependency on Types::Serialiser
Types::Serialiser has a dependency on common::sense, which currently
doesn't contain architecture-specific code but (as of version 3.7)
reserves the right to in a future release; it's already installed into
an architecture-specific subdirectory in PERL5LIB. If such code were to
be added to common::sense, this would break Net::ACME2's "pure Perl or
core dependencies" invariant.

Types::Serialiser is only used once, and this occurrence can be replaced
with a call to an equivalent function that already exists in JSON.pm.
This removes Net::ACME2's dependency on Types::Serialiser, and therefore
common::sense.